backend/requirements.txt
flask>=3.0.0
flask-cors>=4.0.0
tensorflow>=2.20.0
pillow>=11.0.0
opencv-python>=4.8.0
python-multipart>=0.0.6
werkzeug>=3.0.0
python-dotenv>=1.0.0